[REFACT] #176 (#177)
상담 조회, 과제/설문 결과 조회에 대한 api를 어드민 전용 url로 분리
