+++ src/main/java/com/abcdedu_backend/contact/controller/ContactController.java
@@ -1,14 +1,7 @@
 package com.abcdedu_backend.contact.controller;
 
-import com.abcdedu_backend.common.jwt.JwtValidation;
-import com.abcdedu_backend.common.page.PageManager;
-import com.abcdedu_backend.common.page.request.PagingRequest;
-import com.abcdedu_backend.common.page.request.SortRequest;
-import com.abcdedu_backend.common.page.response.PagedResponse;
 import com.abcdedu_backend.contact.service.ContactService;
 import com.abcdedu_backend.contact.dto.request.ContactCreateRequest;
-import com.abcdedu_backend.contact.dto.response.ContactListResponse;
-import com.abcdedu_backend.contact.dto.response.ContactGetResponse;
 import com.abcdedu_backend.utils.Response;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
@@ -17,7 +10,6 @@ import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.data.domain.Page;
 import org.springframework.web.bind.annotation.*;
 
 
@@ -43,26 +35,5 @@ public class ContactController {
         Long contactId = contactService.createContact(contactCreateRequest);
         return Response.success(contactId);
     }
-
-
-    @GetMapping("/")
-    @Operation(summary = "상담 리스트 조회", description = "관리자만 조회 가능 합니다. ")
-    public Response<PagedResponse<ContactListResponse>> readListContact(@JwtValidation Long memberdId, PagingRequest pagingRequest, SortRequest sortRequest) {
-        Page<ContactListResponse> contacts = contactService.readListContact(memberdId, new PageManager(pagingRequest, sortRequest).makePageRequest());
-        return Response.success(PagedResponse.from(contacts));
-    }
-
-    @GetMapping("/{contactId}")
-    @Operation(summary = "상담 조회", description = "관리자만 조회 가능 합니다. ")
-    @ApiResponses(value = {
-            @ApiResponse(responseCode = "403", description = "관리자 전용 기능입니다.", content = @Content),
-            @ApiResponse(responseCode = "404", description = "해당 id에 해당하는 상담을 찾을 수 없습니다.", content = @Content)
-    })
-    public Response<ContactGetResponse> readContact(@JwtValidation Long memberdId, @PathVariable Long contactId) {
-        ContactGetResponse contactResponse = contactService.readContact(contactId, memberdId);
-        return Response.success(contactResponse);
-    }
-
-
 
 }
